markdown_deps: move check_markdown_deps to new module
The function is re-exported under the old name, so this is transparent
to users of the crate.
